filter per selected period
updated array mapping to add casesPerPeriod and deathsPerPeriod to an array returned in switch

cached filterDataByDate value when new dates are set, in order to avoid remapping groupedData, when filters are changed.

added condition check to table, to check whether dates were changed, in order to use groupedData without mapping it to add casesPerPeriod and deathsPerPeriod.
Now table uses totalCases and totalDeaths until dates are changed.